requirements.txt
discord==1.7.3
discord-ext-ipc==2.1.1
discord.py==1.7.3
quart==0.15.1
quart-discord==2.1.4
pymongo >= 3.12.0
python-dotenv>=0.19.0 
dnspython == 2.1.0